Alighed CacheTTL with the duration Terraform output

+++ masherytypes/mashery-v3-types.go
@@ -256,7 +256,7 @@ func (s *Service) Identifier() ServiceIdentifier {
 }
 
 type ServiceCache struct {
-	CacheTtl int `json:"cacheTtl"`
+	CacheTtl int64 `json:"cacheTtl"`
 }
 
 // -----------------------------------------------------------------------------
